+++ src/minesweeper.h
@@ -15,9 +15,9 @@ struct Game {
     int cells_revealed;
     int mine_exploded;
 
-    int cell_size;  // The width/height of each cell in px
-    int x_padding;  // The x offset of the grid in px
-    int y_padding;  // The y offset of the grid in px
+    float cell_size;  // The width/height of each cell in px
+    float x_padding;  // The x offset of the grid in px
+    float y_padding;  // The y offset of the grid in px
 };
 
 int init_game(struct Game *game, int width, int height, int mine_count,
